# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "tmp-backend API",
    "version": "1.0.0",
    "description": "Simple API with register, login (JWT) and /me",
  "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "email": { "type": "string", "format": "email" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "birthday": { "type": "string", "format": "date" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "birthday": { "type": "string", "format": "date" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } }
      }
      ,
      "TransferRequest": {
        "type": "object",
        "required": ["member_code", "amount"],
        "properties": {
          "member_code": { "type": "string" },
          "amount": { "type": "integer", "format": "int64" },
          "memo": { "type": "string" }
        }
      },
      "Transfer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "from_id": { "type": "integer", "format": "int64" },
          "to_id": { "type": "integer", "format": "int64" },
          "amount": { "type": "integer", "format": "int64" },
          "memo": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [],
  "paths": {
    "/register": {
      "post": {
        "tags": ["auth"],
        "operationId": "registerUser",
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Bad Request" },
          "500": { "description": "Server Error" }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["auth"],
        "operationId": "loginUser",
        "summary": "Login and receive JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["auth"],
        "operationId": "getCurrentUser",
        "summary": "Get current user from JWT",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Bearer JWT token, e.g. 'Bearer <token>'"
          }
        ],
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    }
    ,
    "/transfer": {
      "post": {
        "tags": ["transfer"],
        "operationId": "createTransfer",
        "summary": "Create a transfer to a recipient by member code",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Bearer JWT token, e.g. 'Bearer <token>'"
          }
        ],
        "security": [ { "bearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransferRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Transfer" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Recipient not found" }
        }
      }
    },
    "/contacts": {
      "get": {
        "tags": ["transfer"],
        "operationId": "recentContacts",
        "summary": "Get recent transfer contacts",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Bearer JWT token, e.g. 'Bearer <token>'"
          }
        ],
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/users/search": {
      "get": {
        "tags": ["user"],
        "operationId": "searchUserByMemberCode",
        "summary": "Find user by member code (query param 'q')",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Bearer JWT token, e.g. 'Bearer <token>'"
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Member code to search for"
          }
        ],
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      }
    }
  }
}